Generalizartion of OpenSystem
Handling of energies and dipole moments moved to OpenSystem.
